feat(fastapi): use openai instead of langchain
