+++ Dockerfile.dev
@@ -37,7 +37,7 @@ RUN npx puppeteer browsers install chrome --install-deps
 COPY nginx.conf /etc/nginx/nginx.conf
 
 # Expose the port
-EXPOSE 80 3000 8000 6379
+EXPOSE 80 3000 8000
 
 # Start the servers
 CMD ["/bin/bash", "-c", "ollama serve & service nginx start & service redis-server start && node /app/start.js"]
